patterns/agui-strands-agent/requirements.txt
# Strands AG-UI agent dependencies
strands-agents==1.32.0
ag-ui-strands==0.1.1
bedrock-agentcore==1.4.7
mcp==1.26.0
PyJWT[crypto]==2.12.1
